requirements.txt
google-api-python-client==2.119.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gspread==6.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.27
pandas==2.2.1
joblib==1.3.2
tqdm==4.66.2